Refactoring and debugging SIMD perfomances

+++ SpaceshipFactory.h
@@ -5,7 +5,7 @@
 
 class SpaceshipFactory: public Factory{
 public:
-	static Spaceship* get_object(int controller_type, vector<GameObject*>* game_objects, double speed[2], float position[2], int id) {
+	static Spaceship* get_object(int controller_type, vector<GameObject*>* game_objects, float speed[2], float position[2], int id) {
 		int size[2] = { SPACESHIP_WIDTH, SPACESHIP_HEIGHT };
 
 		Controller* spaceship_controller;
